stores: delete the widget-store and the deprecated command aliases
useCurrentStep moves to hooks/use-current-step.ts as pure selectors
over the ViewModel and UiState; the Storybook bypass seeding moves
from the widget-store manual mock to a preview decorator
(seedViewModelBypass in test-fixtures). The empty store shell, its
facade, test, and mock are deleted, and the backend drops the
routing_state_* and permissions_check_all compat aliases together
with the unused permissions.checkAll wrapper. Stage 5 of
docs/2026-06-12-viewmodel-completion-plan.md.
